requirements.txt
asgiref==3.9.1
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
Django==5.2.5
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
iniconfig==2.3.0
mysqlclient==2.2.7
packaging==25.0
pluggy==1.6.0
psycopg2==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
django-filter==25.2
gunicorn==23.0.0